kernelci.org: fix languages config
Fix the configuration and content files hierarchy for multiple
language support.  The default language is set to "en" and all the
content is now moved to content/en accordingly.  Additional languages
may be defined as per the example in the config file.

This also fixes the handy links from the Documentation templates to
GitHub where the pages can be directly edited, as it expects the
language string ("en") to be in the file path.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>