kernelci.org: blog: Add initial blog templates
The design here is to have two categories of blog posts.
1) "News" - For sharing project news
2) "Posts" - For technical, project or any other interesting blog posts

"Posts" was chosen instead of "Blog" as "Blog" has a special meaning and
can only be used as a top-level catch all for all blog syle pages.

Each category has a sub-directory where markdown or html blog posts can be
added.

Initially a couple of dummy posts have been added just so we can see how it
looks. These will be replaced by content from
https://foundation.kernelci.org/news/ and
https://foundation.kernelci.org/blog/.

The structure of how we store the blog post files is yet to be decided.

Signed-off-by: Chris Paterson <chris.paterson2@renesas.com>