kernelci.org: tsc: add section about TSC rules
Add a section to detail the TSC rules in terms of votes and members.
This is based on the project charter document and adds more context as
to how this should work in practice.

A sub-section should be created as a follow-up to archive the results
of all the votes carried out by the TSC.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>