kernelci.org: add Legacy section and update Architecture page
Add a Legacy section to keep documnentation for the legacy system (old
backend, frontend, Jenkins, settings file, YAML config...).  Update
the Architecture page by moving the part about the legacy system to
the legacy section and replacing it with a new description for the
KernelCI API & Pipeline using the up-to-date diagram.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>