kernelci.org: move legacy instances to Legacy section
Move the instances of the legacy system to the Legacy section.

The architecture of the new API means we won't necessarily need to
have the same separate instances as we'll have user accounts to store
data specific to ChromeOS, CIP, Android etc.  We'll still have some
form of staging and production instances though.  This will get
documented as a follow-up and as things get set up.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>