kernelci.org: move legacy maintainers to Legacy/maintainers
Move the entries from the maintainers page that relate to the legacy
system to Legacy/maintainers.  This way, the maintainers page becomes
more up-to-date and we can adjust it with the new API & Pipeline
reployments.  The legacy system is still being maintained but only for
critical issues and not for development any more.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>