docs/admin/api.md: update `Getting an admin token` section
Update documentation for getting an admin token as
per the latest user management changes in the API.
Modify syntax and response for `kci whoami` and
`kci user` commands.
Update documentation to remove details of token
scope as it has been removed and not supported anymore.
Also, it is not possible to pass API token from the
command-line anymore. Tweak the documentation
accordingly.

Signed-off-by: Jeny Sadadia <jeny.sadadia@collabora.com>