kernelci.org: add -F option in docker-compose
Add the Hugo -F option in docker-compose.yaml to also render pages
with a publication date set in the future.

This is useful as the docker-compose setup uses the Hugo development
server for previewing changes, not generating the static files served
in a production deployment where only pages with a publication date in
the past should be rendered.

Signed-off-by: Guillaume Tucker <gtucker@gtucker.io>