kernelci.org: add blog with strategic update about the project
An update to the community in overdue. At the same time, a lot of
project ground shift in direction with the new leadership that in
place.

This article summarize many discussions that happened across
different mediums over the past many weeks.

It is also intended to be a pre Open Source Summit update from
the project.

Signed-off-by: Gustavo Padovan <gustavo.padovan@collabora.com>